+++ src/ProSidebar/MenuItem.tsx
@@ -4,16 +4,17 @@ import classNames from 'classnames';
 export interface Props {
   className?: string;
   icon?: React.ReactNode;
+  active?: boolean;
 }
 
 const MenuItem: React.ForwardRefRenderFunction<unknown, Props> = (
-  { children, className, icon, ...rest },
+  { children, className, icon, active, ...rest },
   ref,
 ) => {
   const menuItemRef: LegacyRef<HTMLLIElement> = (ref as any) || React.createRef<HTMLLIElement>();
 
   return (
-    <li {...rest} ref={menuItemRef} className={classNames('pro-menu-item', className)}>
+    <li {...rest} ref={menuItemRef} className={classNames('pro-menu-item', className, { active })}>
       <div className="pro-inner-item" tabIndex={0} role="button">
         {icon ? (
           <span className="pro-icon-wrapper">